refactor(plus): final refactor for plus page
