+++ app.json
@@ -3,7 +3,7 @@
     "name": "app",
     "slug": "sans",
     "version": "1.0.0",
-    "orientation": "landscape",
+    "orientation": "portrait",
     "icon": "./assets/images/Team_success_Two_Color.png",
     "splash": {
       "image": "./assets/images/Team_success_Two_Color.png",
