+++ src/Views/Plus/AdministrativeHelpSection/index.tsx
@@ -1,6 +1,6 @@
-import SpotlightHelpIllustration from "../Illustrations/SpotlightHelpIllustration";
-import { ImageWrapper } from "../styles";
+import { FlashlightIllustrationWrapper } from "./styles";
 import Text from "~/Components/Ui-kit/Text";
+import SpotlightHelpIllustration from "~/Views/Plus/Illustrations/SpotlightHelpIllustration";
 
 export default function AdministrativeHelpSection(): JSX.Element {
   return (
@@ -9,9 +9,9 @@ export default function AdministrativeHelpSection(): JSX.Element {
         Consulter la liste
       </Text>
       <Text type="titleL">Aides administratives</Text>
-      <ImageWrapper>
+      <FlashlightIllustrationWrapper>
         <SpotlightHelpIllustration width="100%" />
-      </ImageWrapper>
+      </FlashlightIllustrationWrapper>
     </>
   );
 }
