+++ src/Views/Plus/AdministrativeHelpSection/styles.ts
@@ -1,0 +1,10 @@
+import styled from "styled-components/native";
+
+export const FlashlightIllustrationWrapper = styled.View`
+  width: 100%;
+  height: 184px;
+  margin-top: 24px;
+  position: absolute;
+  bottom: -10px;
+  left: -10px;
+`;
