+++ src/Views/Plus/SansaDescriptionNotConnected/styles.ts
@@ -1,0 +1,20 @@
+import styled from "styled-components/native";
+
+export const ButtonsContainerMobile = styled.View`
+  margin-top: 40px;
+`;
+
+export const ButtonsContainerBorne = styled.View`
+  margin-top: 56px;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+`;
+
+export const ButtonWrapperBorne = styled.View`
+  margin-left: 16px;
+`;
+
+export const ButtonWrapperMobile = styled.View`
+  width: 100%;
+`;
